+++ packages/terra-breakpoints/tests/wdio/active-breakpoint-provider-spec.js
@@ -1,8 +1,5 @@
-/* global browser, Terra, before */
-const viewports = Terra.viewports('tiny', 'small', 'medium', 'large', 'huge', 'enormous');
-
-describe('ActiveBreakpointProvider', () => {
-  before(() => browser.url('/#/raw/tests/terra-breakpoints/breakpoints/active-breakpoint-provider-test'));
-  Terra.it.isAccessible({ viewports });
-  Terra.it.matchesScreenshot({ viewports });
-});
+Terra.describeViewports('ActiveBreakpointProvider',
+  ['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], () => {
+    before(() => browser.url('/#/raw/tests/terra-breakpoints/breakpoints/active-breakpoint-provider-test'));
+    Terra.it.validatesElement();
+  });