+++ packages/terra-avatar/CHANGELOG.md
@@ -3,6 +3,23 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added documentation on required peerDependencies
+
+2.20.0  - (July 11, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.19.0 - (July 9, 2019)
+------------------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports
+
+2.18.0 - (June 28, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.17.0 - (June 18, 2019)
 ------------------